refactor:  move classe de request para diretorio
